Modularize arrays
